examples/primitive/requirements.txt
docker==6.1.3
Pillow==10.0.0
typing_extensions==4.7.1
Pillow==10.0.0